refactored controller tests
