+++ app/assets/javascripts/ci_api/controllers/ProactiveChatController.js
@@ -3,7 +3,6 @@ import CommonChatController from './CommonChatController'
 
 export default class ProactiveChatController {
     constructor() {
-        this.sdk = null;
         this.commonChatController = new CommonChatController();
     }
