+++ package.json
@@ -6,6 +6,7 @@
     "@babel/cli": "^7.11.6",
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
+    "@types/jest": "^27.0.2",
     "del": "^2.2.0",
     "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
