+++ test/javascripts/ci_api/controllers/ProactiveChatController.spec.js
@@ -1,0 +1,25 @@
+import CommonChatController from '../../../../app/assets/javascripts/ci_api/controllers/CommonChatController'
+import ProactiveChatController from '../../../../app/assets/javascripts/ci_api/controllers/ProactiveChatController'
+
+describe("ProactiveChatController", () => {
+    it("launches a proactive chat", () => {
+        const commonChatController = new CommonChatController();
+        const proactiveChatController = new ProactiveChatController();
+        
+        const sdk = {
+            isChatInProgress: jest.fn().mockReturnValue(false),
+            getOpenerScripts: jest.fn().mockReturnValue(null),
+            chatDisplayed: jest.fn()
+        };
+
+        window.Inq = {
+            SDK: sdk
+        };
+
+        commonChatController.nuanceFrameworkLoaded(window);
+        proactiveChatController.launchProactiveChat();
+
+        expect(sdk.getOpenerScripts).toHaveBeenCalled();
+        expect(sdk.chatDisplayed).toHaveBeenCalled();
+    })
+});
